use multiplanar format when passing buffers to wayland
MSM stores video data as a single plane so we need to compute the plane
offsets in the buffer to be able to pass the planes separately to the
wayland compositor.

Only support linear NV12 as multiple planes right now as the driver does
not provide enough information to compute the UBWC formats plane
offsets.